мои подписки не работают,изменяла сериализатор и вьюху, почистить #

+++ backend/recipes/models.py
@@ -166,7 +166,7 @@ class RecipeIngredient(models.Model):
         related_name='recipeingredients',
         verbose_name='Ингредиент'
     )
-    amount = models.PositiveIntegerField(
+    amount = models.PositiveSmallIntegerField(
         'Количество',
         blank=False,
         null=False,
